deteccion de login

+++ Web/js/usuario.js
@@ -24,6 +24,17 @@ function actualizarDatos(){
 }
 
 function cerrarSesion(){
-    sessionStorage.clear();
-    window.location.href = "home.html";
+    let http = new XMLHttpRequest();
+ 
+    http.onreadystatechange = function(){
+        if(http.readyState == 4 && http.status == 200){
+                sessionStorage.clear();
+                window.location.href = "home.html";
+        }
+    }
+
+    http.open("POST", "http://localhost:8080/PokemonFBM/desconectarUsuario", true);
+    http.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+    http.send("nombre="+sessionStorage.getItem("usuario")+"&&contrasena="+sessionStorage.getItem("contrasena"));
+    
 }
